Refactoring server and Bugfixes
