+++ server/onlyNextVaccination.js
@@ -1,0 +1,6 @@
+module.exports = function onlyNextVaccination(allDueVaccinations) {
+  return allDueVaccinations
+    .map(array => array[0])
+    .filter(item => item != undefined)
+    .sort((a, b) => a.begins - b.begins)
+}
